Improve type safety.
